--HG--
branch : khc.msnp15
extra : convert_revision : 12d02c8d8c4cbee5a1f9c9aade905aca905c531d

+++ finch/libgnt/gntmain.c
@@ -221,7 +221,7 @@ static gboolean
 io_invoke(GIOChannel *source, GIOCondition cond, gpointer null)
 {
 	char keys[256];
-	int rd;
+	gssize rd;
 	char *k;
 	char *cvrt = NULL;
